refactor(worker): move internal/worker/prompt.go → worker/prompt/builder.go
Last file leaving internal/worker; directory removed. Executor in
worker/pool now calls prompt.BuildPrompt. Only remaining worker→root
cross-module coupling is internal/bot.ParseAgentOutput, marked
transitional in the import block.
