+++ worker/prompt/builder.go
@@ -1,4 +1,4 @@
-package worker
+package prompt
 
 import (
 	"fmt"
